Compile with `-Wall` & warning removal

See merge request art-development-team/ART!14

+++ Source/ClassLibraries/Protocols/ArpCoding.h
@@ -207,7 +207,7 @@ ART_MODULE_INTERFACE(ArpCoding)
         ;
 - (void) codeTableBegin
         : (const char *) tableName
-        : (unsigned long *) codeSize
+        : (unsigned int *) codeSize
         ;
 - (void) codeTableEnd
         ;
